split gromov.py in ot/gromov/ + update test_gromov with helper_backend functions
